requirements.txt
Django>=5.1.6
python-dotenv>=1.0.0
psycopg2-binary>=2.9.5
dj-database-url>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
